+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# x.x.x
+
+- fix: fix bug where localizations were not resolved correctly due to nested `MaterialApp`s
+
 # 0.3.3
 
 - chore: upgrade flutter to 2.10.4